fixed api response issue on match making

+++ app/src/main/java/tech/iosd/gemselections/Retrofit/ResponseModels/MaleBirthDetails.java
@@ -9,6 +9,7 @@ import com.google.gson.annotations.SerializedName;
 
 public class MaleBirthDetails {
 
+
     @SerializedName("year")
     @Expose
     private int year;
@@ -26,13 +27,13 @@ public class MaleBirthDetails {
     private int minute;
     @SerializedName("latitude")
     @Expose
-    private int latitude;
+    private float latitude;
     @SerializedName("longitude")
     @Expose
-    private int longitude;
+    private float longitude;
     @SerializedName("timezone")
     @Expose
-    private int timezone;
+    private float timezone;
     @SerializedName("sunrise")
     @Expose
     private String sunrise;
@@ -83,7 +84,7 @@ public class MaleBirthDetails {
         this.minute = minute;
     }
 
-    public int getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
@@ -91,7 +92,7 @@ public class MaleBirthDetails {
         this.latitude = latitude;
     }
 
-    public int getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
 
@@ -99,11 +100,11 @@ public class MaleBirthDetails {
         this.longitude = longitude;
     }
 
-    public int getTimezone() {
+    public float getTimezone() {
         return timezone;
     }
 
-    public void setTimezone(int timezone) {
+    public void setTimezone(float timezone) {
         this.timezone = timezone;
     }
